refactor commons
