+++ python-lib/regression_splines/commons.py
